+++ package.json
@@ -1,5 +1,20 @@
 {
   "dependencies": {
     "decimal.js": "^10.4.3"
-  }
+  },
+  "devDependencies": {
+    "webpack": "^5.91.0",
+    "webpack-cli": "^5.1.4"
+  },
+  "name": "calculator",
+  "version": "1.0.0",
+  "description": "This is a apple calculator on web made with HTML, CSS, Vanilla Javascript",
+  "main": "index.js",
+  "scripts": {
+    "build": "webpack",
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC"
 }
