+++ public/index.html
@@ -7,8 +7,8 @@
 	<meta name="description" content="A website of iOS built-in Calculator" />
 	<meta name="keywords" content="iOS, Calculator, iOS Calculator, iOS Calculator App, iOS" />
 	<title>Calculator</title>
-	<link href="style.css" rel="stylesheet" type="text/css" />
-	<link href="align.css" rel="stylesheet" type="text/css" />
+	<link href="../src/style.css" rel="stylesheet" type="text/css" />
+	<link href="../src/align.css" rel="stylesheet" type="text/css" />
 </head>
 
 <body>
@@ -50,8 +50,10 @@
 		</div>
 	</div>
 	<!-- Calculator -->
-	<script src='path/to/decimal.js'></script>
-	<script type="module" src="test.js"></script>
+	<!-- <script type="module" src="../dist/main.js"></script> -->
+	<!-- <script src='path/to/decimal.js'></script> -->
+	
+	<script type="module" src="../dist/main.js"></script>
 	<!-- <script src="script.js"></script> -->
 </body>
