+++ webpack.config.js
@@ -1,0 +1,9 @@
+const path = require("path");
+
+module.exports = {
+    entry: './src/index.js',
+    output: {
+        filename: 'main.js',
+        path: path.resolve(__dirname, 'dist')
+    }
+}
